# specs/003-app-integration/contracts/openapi.yaml
# OpenAPI Specification: App Integration (Todos)
openapi: 3.0.3
info:
  title: Todo Web App API
  version: 1.0.0
  description: Secure API for managing user-specific todo items.
servers:
  - url: http://localhost:8000/api/v1
    description: Development backend

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Todo:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 500
        completed:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - title

    TodoCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - title

    Error:
      type: object
      properties:
        detail:
          type: string

security:
  - bearerAuth: []

paths:
  /todos:
    get:
      summary: List all todos for the current user
      responses:
        '200':
          description: A list of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get a specific todo
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Not found

    patch:
      summary: Update a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

    delete:
      summary: Delete a todo
      responses:
        '204':
          description: Deleted successfully